17311: Make region required based on selected country
Before this commit the region was required for every country, it did not
validate the setting 'State is required for' under 'Stores >
Configuration > General > General > State Options'.

By using the `getCountriesWithStatesRequired` method from the already
through DI available `Magento\Directory\Helper\Data` class we can define
if for the current selected country the region is required.

Removed checks if elements contain a class before removing it. It does
not matter if an element has the class or not. Removing the checks saves
some Javascript time.

Also fixed:
+ Imported classes
+ Optimize if/else statement for determining if country is set