Fix allowing SQL injections in addExpressionFieldToSelect
The str_replace function used in `addExpressionFieldToSelect` directly adds the string in `$fieldItem` to the query. Here `$expression` must be checked by the developer to not have any dangerous content. But as developers expect magento core functions to prevent sql injection, they are prone to put input parameters directly inside `$fields` which can lead to sql injections. Other function `addFieldToSelect` in the same class ends up with quoted text. But this is not the case with this function.