Fix issue with possibly colliding form-fields
It is possible for modules to extend the form-fields within a grid. Therefore it can happen that other formfields - especially hidden fields - are added to a grid so that the query does not only return checkboxes but also other form-fields. And then the first form-field is not necessary the checkbox we want in this case.

This change alters the selector that is used to query for the first checkbox by actually only requesting checkbox-fields. So the chances are much higher that the checkbox-field we want is actually the first item of the list.