When setting `background` for labels explicitly the labels in admin will overlap the `::before` pseudo-element.
I suggest to remove explicit setting for `label` element or replace
`@color-white` to `transparent`.

In some cases developers try to hide elements by setting white
background and higher z-index for the element with white background.
But the proper way of hiding elements is to override templates config or
remove elements by JavaScript.