Patch #22309: Remove "null" value of not passed name attribute
If a category is updated via REST API in the "all" scope and the
category name is not sent, Magento use "null" as value for the name,
This leads into a validation error, because the category name is
required.
In this case we remove the null value out of the automatically
generated default value array. Magento will then skip the category,
because there is no change.