Fixed issue with printing in firefox
Tested with order print page in Firefox and problem of only printing the first page is solved. Maybe also fixes print problems with other pages.
The `display` style for `.page-wrapper` and `.columns` is by default set to `flex`. This pull request sets them to `block` for printing therefore fixes the problem with printing in Firefox.