Do an empty check instead of isset check on image removed
In some cases the 'removed' tag on the image is already set to an empty
string in the data which will be loaded. Therefore the isset check fails
which clears the image roles.

On all other cases where this 'removed' tag is being validated with an
empty or !empty. Therefor it seems safe to apply this here also.