Fix the 404 message for product alerts when not logged in
When the user isn't logged in and presses the product alert link
(either stock or price), the user has to login and gets redirected
to a 404 page. This is because the redirect done by the login page
is a `GET` call. The fix done is to change the used interface for
the product alerts from `POST` to `GET`, as this is the easy fix and
there's no reason to use `POST` here at this moment, as all parameters
for the call are all in get URL parameters.