Fix missing attribute_id condition from filter
This commit adds a missing attribute_id condition to the
addAttributeToFilterAllStores method in product dataprovider.

Previously the method did not include the attribute_id as a condition.
This meant that it was impossible to filter by status as any value
of 1 or 2 in the catalog_product_entity_int table would match the
condition.