Generate nonce from public hash server-side
This simplifies the implementation requirements when setting the payment
method to Braintree Vault.

In the current frontend implementation the nonce is created server-side, but
using an ajax request when the place order button is activated. The current
storefront flow executes:
1. User selects vault method and places order
2. Ajax request with public hash is made to server returning the nonce
3. Nonce is submitted with order placement request in payment details

The second step is uncessecary and is removed from the graphql flow with this
commit.