Fix getCurrentUrl() not respecting store base path
This is correcting an issue for magento where magentos
\Magento\Framework\Url::getCurrentUrl()
function is not respecting the stores base url.

----
Background:

In setups, where the stores are
switched with the setting of
  SetEnv MAGE_RUN_TYPE store
  SetEnv MAGE_RUN_CODE de_de
the path that is available in the php request object is in our setup
not including the complete path but only the path after the
storeBaseUrl.

 url in browser: https://example.com/de/shop/hosen/
 host in php: example.com
 path in php: /hosen/
 base url in Magento: https://example.com/de/shop/

Now we running into an issue with the add to compare list links. This
links are generated and the redirect url is set already in the link
itself with the 'uenc' parameter. And this parameter contains a wrong
url: https://example.com/hosen/

After tracing down where this wrong url is comming from, it looks like
in the getCurrentUrl function of magento the base url of the shop is not
respected. Therefor I decided to change this function in a way that is
respecting the store base url if it is set and falls back to server/php
url otherwise.

Additionaly this also refector the url generation in a way, that is not
relying on manualy determine the parts of the url, but using the
zend/php framework for this.

Also updates the class constructor of Magento/Backend/Model/Url as it is
extending Magento/Framework/Url and would call the parrent constructur
in a wrong way after including the store manager into di.