Clone product before adding to cart
This prevents undesired behavior when adding multiple variants of the same SKU
to the cart in a single request. Without cloning the product when adding to,
the product object set with the quote item is a referenced object. When adding
a n+1 cart item with same sku, but different options multiple cart items are
referencing the same instance of a product object. This causes
Item::representProduct to ignore custom option and variant differences.

Fixes magento/graphql-ce#854