Fix GraphQL schema @cache declaration argument typo
The `@cache` directive argument name is `cacheable`, not `cachable`.

Reference line 1 of the method `\Magento\GraphQlCache\Model\CacheableQueryHandler::handleCacheFromResolverResponse`:

```
public function handleCacheFromResolverResponse(array $resolvedValue, array $cacheAnnotation) : void
{
    $cacheable = $cacheAnnotation['cacheable'] ?? true;
    $cacheIdentityClass = $cacheAnnotation['cacheIdentity'] ?? '';

    if ($this->request instanceof Http && $this->request->isGet() && !empty($cacheIdentityClass)) {
        $cacheIdentity = $this->identityPool->get($cacheIdentityClass);
        $cacheTags = $cacheIdentity->getIdentities($resolvedValue);
        $this->cacheableQuery->addCacheTags($cacheTags);
    } else {
        $cacheable = false;
    }
    $this->setCacheValidity($cacheable);
}
```

In the context of Magento, the directive declaration is only for documentation purposes.
The directive itself is used by the GraphQL API implementation, but the directive declaration is not.
As such this PR is backward compatible. The effect is only that, once merged, valid directive are no longer highlighted as an error in PHPStorm.