Fix label being cutoff in Newsletter registration
In this commit the label to Register in the newsletter that exists in the homepage had the text completely cut off on Tablet iOS devices

To solve this the width property was altered to use the max-content value, and a max-width property is also added. Also some paddings were set to 0 because Safari keeps adding some default padding values.