Fixes the problem with page reload when js prompt was closed with enter
Forms inside prompt.js did not have any submit action, so hitting enter
inside any of the inputs caused submission to the current page - reload of the page.
All the JS logic was lost.
I added a default submit handler (whenever there is a form in the content of the prompt)
and default action that is the same as pressing the OK button on the prompt.
If one does not want this behavior, he can pass an empty selector in options,
and it will disable the onSubmit listener.
This update will prevent faulty submission (by hitting Enter key) on all prompts.