Adds color conversion to SRGB for images with CMYK and Unrecognizable color palette
For ImageMagic, when the original image was in CMYK palette, the final image after
resize was kind of "in negative". I decided to force conversion to RGB, but when
converting to RGB from CMYK with IMagixk - colors were not natural.
So I picked SRGB (I found this hint in StackOverflow)