Fix 'promotion' region not rendering
The getRegion(name) method returns a knockoutjs observable
array which is a function, so the check used in the template
will always resolve to false (as function `length` property
is the number of arguments a function takes, which for
the observable array is 0). Therefore the region is never
rendered.

The fix from this PR is a bit ugly, but simple.