Prevent page scroll jumping when product gallery loads
During the investigation of the issue linked below I was able to pinpoint the exact moment of scroll jumping to fotorama's `that.resize` function
which for some reason resets all of its major elements' width and height dimensions. I analysed the code and concluded that since current element
height seems to be completely ignored in the whole script and is only set either based on settings or its width and ratio, the code that resets it
can be removed. Please not that this can be also true for width but I left it here just to be safe as it doesn't contribute to page jumping.

In addition, `min-height` was added to prevent initial jump of the gallery when static image is being replaced by fotorama's HTML.

Fixes https://github.com/magento/magento2/issues/10518.