Update composer dependency to fix Redis Key Expiery
I experienced a situation on production system with 15 stores and
organic trafic, where the redis cache grows to over 30GB. It looks like
magento is putting a lot of entries into the redis cache without any
expiary. I needed to limit redis memory to prevent the server from
dying, not how it should be. This had some performance influences.

For this issue there is actualy a fix in place which just need to be
pulled in. This commit is do this. With the fix in place the redis
stabalized at arround 3GB. Over all performance is was also increased
with this change.

Referces:
[1] https://github.com/colinmollenhour/Cm_Cache_Backend_Redis/commit/bc63e7226a525fe95865c756aad0037f0d57d03c
[2] https://github.com/magento/magento2/commit/32058c748731348e7d38cffa7a6c56113e17b3d8
[3] https://github.com/magento/magento2/issues/25487