Return HTTP response of example import file as CSV
We are return content of CSV file here. So without these changes unnecessary file is created in "var" directory and default response type is returned. So file is not downloaded, just displayed in browser. This fix will allow to actually download example file and not use some "hacks" in other modules.