Fix date not being saved from the admin when the locale is a different language than english
The issue was that for some reason when saving a product entity and a date was added, the time was also calculate and it was always set to 0:00am. There is no point for a calculation like that.

By sending the value false as a last parameter during the call of ->date function, we allow magento to process a date without adding the dummy 0:00am in the datetime that is causing the whole saving process to fail.