Issue #1054: Add e-mail address validation to customer data validation for GraphQL
The goal is to make customer e-mail address validation in accordance to the validation in Admin panel.

Currently, there are 2 validation on customer create/edit page in Admin:
1) frontend validation (see 'validate-email' in app/code/Magento/Ui/view/base/web/js/lib/validation/rules.js),
2) backend validation (see '_validateInputRule' in www/app/code/Magento/Eav/Model/Attribute/Data/AbstractData.php).

In order to make validation in GraphQL working the same, the check using regular expression from JS validation was used as a first step and original Zend validation as a second.