Removes setting domain when domain for cookies empty
The previous implementation set the domain for cookies to the current
domain provided in the Request object.
As per RFC - https://tools.ietf.org/html/rfc6265#section-4.1.2.3,
this is the default behavior to use the current domain (and only
current domain) when the domain is not set in setcookie call.
Setting the domain explicitly during setcookie has a side effect
- the cookie is also visible for subdomains.
This visibility is probably not intended.  Current implementation
(before my change) is based on old  RFC (https://tools.ietf.org/html/rfc2109)
when domain shall be prepended with dot to make it visible for subdomains.
If a user wants to include cookie for subdomains - shall provide
the domain in cookie settings.

I also modified tests to reflect this new logic. Tests were failing only
because it was checking the execution of the getHost method from requests.
That method is no longer needed/used.