Allow wishlist share when all items are out of stock
Clicking `Share` on the wishlist view page updates and saves the wishlist
before redirecting to the share page. The redirection to the share page only
happens when the post body includes wishlist descriptions. Because qty and
comment are disabled for out of stock wishlist items these properties are not
sent with the post resulting in the share redirect being unreachable when all
items are out of stock.

This commit updates the controller flow separating the logic conditions for
saving and sharing. This allows sharing wishlist where all items are out of
stock.