Remove Filename Normalization in Delete Controller
Filenames are normalized when saving to the server via the uploader.

https://github.com/magento/magento2/blob/241271e8417c4264d44682169aa2032e955d6942/lib/internal/Magento/Framework/File/Uploader.php#L391-L407

This normalization was added to the admin delete controller in 2.2.0

https://github.com/magento/magento2/commit/09d662e2a163049d7d09c8e23e60a547a4b0400a#diff-7c65d1bd4c41efed1d26ddf72f15aa91R62

This change prevents admin users from deleting CMS images not conforming to the
uploader normalization. For instance an image on the server named ` - .jpg`
would have it's name normalized to ` _ .jpg` in the delete controller which
causes the deletion to fail.

Fixes magento/adobe-stock-integration#889