Fix: missing "null" and "false" return type hints in RouterInterface::match()
The core code in FrontController expects a falsy value to be
returned from routers that don't match a request, but want to
delegate to the next router. The code in Webapi\Rest\Router expects
a false in that same case.

However those concrete expectations are not reflected in the
RouterInterface::match return type hint (DocBlock) and are not
documented either.

This commit:
* Adds the "null" and "false" return type hints to the RouterInterface::match()
docblock. This fixes issues with PHPStan and IDE's, and makes the behaviour more
explicit.
* Adds a unit test to FrontControllerTest to make sure specifically returning "null"
works as expected in FrontController.