Fixed URL Rewrite generation and removal on product/website change.
Moved scope of ProductToWebsiteChangeObserver to base instead of adminhtml
as this functionality has been moved to a consumer. With the observer set
to adminhtml, this observer was not being called properly on changing
the websites for a particular product in the consumer.

Corrected ProductProcessUrlRewriteSavingObserver to correctly add and remove
rewrite URLs when products are added and removed from websites using the
Product Edit Admin Page.

Corrected ProductToWebsiteChangeObserver to correctly add and remove
rewrite URLs when products are added and removed using the Products
Grid Mass Action function.

Extended unit tests to cover both Observers.

Fixed MFTF test for Product URL Rewrite generation.