Fix for Issue #26284
Fix for Issue: Incorrect Tax showing for online Credit Memo #26284
Link: https://github.com/magento/magento2/issues/26284

The issue was caused by the fact that the $allowedTax and $allowedBaseTax variables were set to the same values without taking into consideration if the current creditmemo is being created for an invoice. The commit makes sure to subtract the refunded tax for any credit memo that is associated with the current invoice.

Ideally a getCreditmemosCollection(), getTaxRefunded() and getBaseTaxRefunded() functions could also be integrated directly into the Magento\Sales\Model\Order\Invoice class.