Convert MSRP currency of configurable product options
The MSRP of configurable products is updated in Javascript when an
option is chosen. Javascript uses an object with prices of the options.
The prices in this object are converted to the chosen currency. MSRP
prices however were not converted, leading to the price being displayed
in a wrong currency.