Use helper method that doesn't throw when no model
When a payment configuration node exists in XML but there is no <model>
defined, getList() will throw an UnexpectedValueException. One use case for
creating such a set-up is a module that provides default configuration
overrides for payment methods which are not installed on the current website.
For example:

<config>
  <default>
    <payment>
      <checkmo>
        <active>0</active>
      </checkmo>
      <not_installed_here>
        <debug>0</debug>
        <environment>production</environment>
      </not_installed_here>
    </payment>
  </default>
</config>

In this case, the payment method 'not_installed_here' does not have a <model>
node defined, as the payment method module is not installed on the website.