Fix #29879 Breadcrump Undefined class constant 'XML_PATH_CATEGORY_URL_SUFFIX'
When creating a plugin, that is calling the public functions for the
product breadcrump they ending up throwing an error message, because
this constants has been marked as private recently.

This is fixing it with referencing to the current instance and all is
working again. This fix is analog to another bug of the same type:

https://github.com/magento/magento2/issues/28981
https://github.com/magento/magento2/pull/28797

This should resolve:
https://github.com/magento/magento2/issues/29879