Fix array to string conversion error when saving row system config with defaults
This PR solves issue #30314. The backend model that is designed to handle complex values serializes non-scalar values before they are persisted to database. But when the old value is loaded from config defaults (defined in config.xml) for comparison, the default value is fetched as array. When the array is casted to string it results in an array to string conversion error.

This issue is fixed by serializing the default value coming from the config.