Set cartItem errors relative to the cart item
Currently the errors are always returned at the beginning of the cart items array. 
This makes it impossible for the frontend to know error is associated with a cart item because the error paths are returned incorrectly.

It would be best to fix the cart so it won't return additional `null` items.