Use datepicker Knockout binding to achieve two-way value sync in date field
The previous method implemented in the commits linked below was broken,
because the jQuery datepicker initializes it's value to the current date.
Since it was using the `shiftedValue` property of the date field,
all empty date fields would get initialized with the current date on page load.

This commit replaces the fix implemented in the commits below.
- https://github.com/magento/magento2/pull/31605/commits/1d9e1b6fb9b45a3036bb729cd2841661b50a30c0
- https://github.com/magento/magento2/pull/31605/commits/41b4e17559105b03b8c76c7151713e000582f880