Fix loading of calendar script
There is a bug with the loading of mixins in Safari and Firefox. 
When you call the widget using the imperative notation: ".calendar({", the mixin for this module is not loading in Safari and FIrefox. 
We should use declarative notation to initialize the widget to avoid this problem: <script type="text/x-magento-init">