Use the cookie_path config to set the messages cookie
This is a fix to take into account the cookie path that has been configured in the magento backend when setting the mage-messages cookie.
If this is not set, there is an inconsistency between the PHP and JS where the lib/web/mage/cookies.js takes into account the path and the backend PHP code doens't.