Update di.xml
The default implementation of the `Magento\Framework\HTTP\ClientInterface` is `Magento\Framework\HTTP\Client\Curl` which is a stateful class

When a `Magento\Framework\HTTP\Client\CurlFctory` is injected, there are no side effects because instances are not shared.

But when an instance of `Magento\Framework\HTTP\ClientInterface` is required (e.g., in the `Magento_InventoryDistanceBasedSourceSelection` module) through constructor-based dependency injection, this could have side effects, because a shared instance of a stateful class is used. If, for example, you set a curl timeout on that shared instance, that timeout will affect every class that uses the same instance.