Update scripts.js
Figured out the correct fix based on other methods. The returned object from this.serializeData should then have toObject() call on it so that data is an object of the serialized data. This is how all other instances of this.serializeData work in this file and changing this fixes the issue.