Filter only visible and enabled products
In the VisibilityStatusProcessor there was a filter for status and
visibility using `joinAttribute` this resulted in getting all
products where `status` was greater than 0 and `visibility` was
greater than 0. However, the "disabled" status in Magento is `2`
and the "not visibile" visibility in Magento is `1`, resulting in
all products, including the non-visibile and disabled products,
where returned in the GraphQL results.