Update CONTRIBUTING.md
Adding correct path for Forks And Pull Requests Instructions. Added updated link for beginners Slack chat, Which will redirect the user to the Slack Application instead of opening the Slack Channel in browser. If the user is not having a Slack Application, then beginners Slack chat will open in the browser.