Enforce array type in DesignTheme config processor
Without this, an empty config node like `<my_store />` in a `config.xml` breaks installation, as `null` is passed to the private method.

Since it is private this change is 100% backwards compatible. I'm not adding a test case since not having regressions should be a sufficient merge criteria when adding types.