Add --exclude-group to cron:run
https://devdocs.magento.com/guides/v2.4/config-guide/cli/config-cli-subcommands-cron.html#config-cli-cron-group-run

When you decide to split a single cron group with the `--group` flag it means you need to do so for every group in your instance.

It's pretty common for third party modules to define their own groups whenever they like, so we have to keep up to date with the defined groups in the instance to keep adding these to our crontab when required.

One of my instances currently has 11 groups to manage, more will come in the future.

In my case I only really want to run the crons like

```
* * * * * bin/magento cron:run --exclude-group index
* * * * * flock -n index.lock bin/magento cron:run --group index
```

This will give me better granular control over the indexer crons which can be useful when your store reaches hundreds of thousands of SKUs

For future proofing I've made the `--exclude-group` flag stackable so you can define it multiple times like
```
* * * * * bin/magento cron:run --exclude-group index --exclude-group consumers
```