Update indexer:reset to call indexer->invalidate()
I was recently debugging something that was invalidating the indexers. 

I added some logging using a before plugin on the `invalidate` function

https://github.com/magento/magento2/blob/227cf9229152ed4332c607796cf476055ef9ec4f/app/code/Magento/Indexer/Model/Indexer.php#L345-L355

But this use case isn't covered. It seems to me that this command should just call the function that invalidates the indexer rather than repeating the work itself.