Reset custom attribute value after it is updated with setData
This solves an issue where an old custom attribute value becomes stale in the
custom attributes array after the value of the custom attribute is changed
through setData instead of setCustomAttribute. The fix is to remove the value
from the custom attributes array in setData, which causes the value to be
updated to the custom attributes array the next time custom attributes are
accessed (due to setData setting the customAttributesChanged property as
enabled).

A concrete example of this issue (and what prompted this patch) is the
\Magento\Eav\Model\Entity\Attribute\Backend\ArrayBackend backend model for
multiselect attributes. Without this patch, a custom multiselect attribute's
value is not saved. The backend model does map the array value correctly
into a comma-separated string, and sets it to the object using setData, but it
does not update the attribute value to the custom attributes array. This is a
problem, because Magento ends up overwriting the correct comma-separated value
from the stale custom attributes array when extracting data to save in
\Magento\Framework\EntityManager\Operation\Update\UpdateMain::execute.

This is identical to what is done in unsetData in regards to custom
attributes.