Make sure $key is not type juggled when comparing to string
When saving a dynamicRows in a category (for example), the array gets passed through this function.
The first element of this array has $key = 0; When PHP tries to compare it to the 'custom_attributes' string, it casts the string to a number, resulting in 0. This causes filterCustomAttributes to be called, which in turn calls flattenCustomAttributesArrayToMap. This raises a exception since it expects a array, and we are passing in a string.