Make admin abstract field array -widget html ids unique.
The configuration data can be stored in the core_config_data -table as either array or object.
If the data is stored as an array, the table and it's rows will use array index numbers
as html ids. If the configuration page has multiple of these widgets, there will be duplicate
ids on the page, which will cause javascript issues when the scripts try to read/update rows
since the rows ids are not unique. Deletion of rows will also not work.