Prepare abstract field array rows by making them valid for rendering.
If a row is missing one of it's column data, the whole configuration page will
break due to a javascript error when the rows are being rendered. This commit will
make sure that all of the rows are rendered. If a column is a dropdown, the validity doesn't
need to be checked since dropdowns will be handled by their renderers.