Ensure custom attributes don't get mangled in AbstractAddress
If Abstract Address is called with `->setCustomAttributes()` and is given an array of Attribute objects, it will convert them into an array.  When it does this, it turns the array from something like `['attribute' => Object]` to `['attribute' => ['attribute_code' => 'attribute', 'value' => 'value']]`

This eventually gets sent back through to AbstractExtensibleModel's `filterCustomAttributes` which sees it as an array, but does not see it as an indexed array, and so does not `flattECustomAttributesArrayToMap`, resulting in a Attribute object being made where the value is `['attribute_code' => 'attribute', 'value' => 'value']`.