Lower retrieval rate of disabled module output
This change ensures we retrieve the modules_disable_output data only once and use that for every block output where we need to check if the output is disabled. This improves performance by a bit as well. Only pages that have all cache enabled (excl. FPC) the gains are 5-10m depending on the page, this amount is bigger on pages that is missing block_cache for certain blocks, 30ms

Also the get for getHtml has been changed to getData('html'). There is a minimal improvement to performance, without block cache (it was hit 1700 times) on a PLP the gain is about 6-7ms, and with full block_cache its only 1ms, but it seems like an easy gain.