Add caching for route urls to allow for more cache collisions
This change will improve route url retrieval in combination with different variants of a url with query params and ensure query params are not taking into account when generating a cached entry. 

On a PLP we have cases where this method is called 300-900 times depending if there is block_cache or not:
For a fully cached page excluding FPC, this is called 327 times and takes 31ms and with patch 12ms.
For a fully cached page excluding FPC and Block_cache, this is called 836 times and takes 38ms and with patch 12.5ms.

The gains can be much bigger depending on the situation. For another client we have seen bigger gains where more variable urls within certain facets were given with query params, resulting in a bigger impact.

A different solution might be proposed of course, but I think the potential gain should be looked at.